fix PR #14
