+++ src/containers/SelectToken/index.tsx
@@ -24,7 +24,7 @@ interface selectTokenProps {
   onChange: (_: SelectTokenType) => void;
 }
 
-const SelectTokenContainer = ({ onChange }: selectTokenProps) => {
+const SelectToken = ({ onChange }: selectTokenProps) => {
   const [selectedToken, setSelectedToken] = useState<null | Horizon.BalanceLine>(null);
   const [isOpen, setIsOpen] = useState(false);
   const [searchValue, setSearchValue] = useState('');
@@ -46,8 +46,10 @@ const SelectTokenContainer = ({ onChange }: selectTokenProps) => {
     setIsOpen(true);
   };
 
-  const filteredOptions = userData.filter((option) =>
-    option.asset_code?.toLowerCase().startsWith(searchValue.toLowerCase()),
+  const filteredOptions = userData.filter(
+    (option) =>
+      option.asset_type === 'native' ||
+      option.asset_code?.toLowerCase().startsWith(searchValue.toLowerCase()),
   );
 
   const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -70,7 +72,7 @@ const SelectTokenContainer = ({ onChange }: selectTokenProps) => {
               height={35}
               alt=""
             />
-            <p className="ml-4 text-midnightBlue">{selectedToken.asset_code}</p>
+            <p className="ml-4 text-midnightBlue">{selectedToken.asset_code || 'XLM'}</p>
           </div>
         ) : (
           'Select token'
@@ -88,7 +90,7 @@ const SelectTokenContainer = ({ onChange }: selectTokenProps) => {
           {filteredOptions.map((i) => (
             <div
               className="flex items-center w-full cursor-pointer h-[72px] border-b last:border-none"
-              key={i}
+              key={i.asset_code}
               onClick={() => handleTokenSelect(i)}
             >
               <div className="flex w-full items-center">
@@ -103,7 +105,9 @@ const SelectTokenContainer = ({ onChange }: selectTokenProps) => {
                   />
                 </div>
                 <div className="text-left w-full">
-                  <p className="text-black text-base w-full font-bold">{i.asset_code}</p>
+                  <p className="text-black text-base w-full font-bold">
+                    {i.asset_code || 'XLM'}
+                  </p>
                 </div>
               </div>
 
@@ -121,4 +125,4 @@ const SelectTokenContainer = ({ onChange }: selectTokenProps) => {
   );
 };
 
-export default SelectTokenContainer;
+export default SelectToken;
