+++ src/containers/Summary/index.tsx
@@ -10,7 +10,7 @@ import { FormValues } from '../CreateStreamMainCard';
 import { rateInNumber } from 'src/utils/rateInNumber';
 import { calculateTotalAmount } from 'src/utils/calculateTotalAmount';
 import { checkBalance } from 'src/utils/checkBalance';
-import { getFormValues } from './formValues';
+import { mapFormValues } from './mapFormValues';
 
 import summaryLogo from 'public/images/summary.svg';
 
@@ -21,6 +21,7 @@ interface SummaryProps {
 
 const Summary = ({ form, isFormValidated }: SummaryProps) => {
   const values: FormValues = form.getValues();
+  const getFormValues = mapFormValues(values);
 
   let totalAmount = new BN(0);
   let errorMessage;
@@ -53,8 +54,8 @@ const Summary = ({ form, isFormValidated }: SummaryProps) => {
     <div>
       <CPageCard title={summaryTitle} className="px-3 py-4 mb-4 w-[80%] ">
         <ul className="grid gap-2 text-midnightBlue">
-          {getFormValues(values).length > 1 &&
-            getFormValues(values).map((x) => (
+          {getFormValues.length > 1 &&
+            getFormValues.map((x) => (
               <li
                 key={x.label}
                 className="flex justify-between w-full whitespace-nowrap overflow-hidden text-clip items-center bg-alabaster h-10 px-4 text-sm rounded-[10px]"
