+++ src/stories/SelectTokenContainer.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import SelectTokenContainer from '../containers/SelectTokenContainer';
+import SelectTokenContainer from '../containers/SelectToken';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
